requirements.txt
langchain-community==0.0.6
langchain-core==0.1.1
langchain-openai==0.0.3
python-dotenv==1.0.0
openai==1.3.5
pypdf2==3.0.1
scholarly==1.7.11
habanero==1.2.3
nltk==3.8.1
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
pdfplumber==0.10.3
pdfminer.six==20221105
